LayoutTests/platform/chromium/fast/inspector-support/uncaught-dom3-exception-expected.txt

CONSOLE MESSAGE: line 9: Uncaught Error: HierarchyRequestError: DOM Exception 3
This tests that we are getting the correct message for DOM Exception 3: HIERARCHY_REQUEST_ERR.
